refactoring json protocols
