+++ smrt-common-models/src/main/scala/com/pacbio/common/models/JsonProtocols.scala
@@ -2,6 +2,7 @@ package com.pacbio.common.models
 
 import java.util.UUID
 import spray.json._
+import org.joda.time.{DateTime => JodaDateTime}
 import fommil.sjs.FamilyFormats
 
 trait UUIDJsonProtocol extends DefaultJsonProtocol with FamilyFormats {
@@ -11,6 +12,20 @@ trait UUIDJsonProtocol extends DefaultJsonProtocol with FamilyFormats {
     def read(json: JsValue): UUID = json match {
       case JsString(x) => UUID.fromString(x)
       case _ => deserializationError("Expected UUID as JsString")
+    }
+  }
+}
+
+trait JodaDateTimeProtocol extends DefaultJsonProtocol with FamilyFormats {
+  import PacBioDateTimeFormat.DATE_TIME_FORMAT
+
+  implicit object JodaDateTimeFormat extends JsonFormat[JodaDateTime] {
+    def write(obj: JodaDateTime): JsValue = JsString(obj.toString(DATE_TIME_FORMAT))
+
+
+    def read(json: JsValue): JodaDateTime = json match {
+      case JsString(x) => JodaDateTime.parse(x, DATE_TIME_FORMAT)
+      case _ => deserializationError("Expected DateTime as JsString")
     }
   }
 }
