+++ smrt-server-base/src/main/scala/com/pacbio/common/actors/CleanupDao.scala
@@ -11,7 +11,7 @@ import com.pacbio.common.dependency.Singleton
 import com.pacbio.common.logging.{LogResources, Logger, LoggerFactory, LoggerFactoryProvider}
 import com.pacbio.common.models._
 import com.pacbio.common.services.PacBioServiceErrors
-import com.pacbio.common.time.{Clock, ClockProvider, PacBioDateTimeFormat}
+import com.pacbio.common.time.{Clock, ClockProvider}
 import com.pacbio.secondary.analysis.engine.CommonMessages.MessageResponse
 import org.joda.time.{DateTime => JodaDateTime, Duration => JodaDuration}
