+++ smrt-server-link/src/main/scala/com/pacbio/secondary/smrtlink/models/DataModelParser.scala
@@ -8,7 +8,7 @@ import javax.xml.datatype.XMLGregorianCalendar
 
 import com.pacbio.common.dependency.Singleton
 import com.pacbio.common.services.PacBioServiceErrors.UnprocessableEntityError
-import com.pacbio.common.time.PacBioDateTimeFormat
+import com.pacbio.common.models.PacBioDateTimeFormat
 import com.pacificbiosciences.pacbiobasedatamodel.SupportedAcquisitionStates
 import com.pacificbiosciences.pacbiodatamodel.PacBioDataModel
 import org.joda.time.{DateTime => JodaDateTime}
